Cluster-density refactor
The current cluster-density workload is outdated, and not very
representative of a real case scenario. It makes an extensive usage of
builds, the extensive usage of this object generates lot of delays in
the workload because of long building times and can also lead to timeouts.
Also, generating such amount of build objects is not very representative
of a real case scenario.

Replacing most of these builds by deployments and more services will generate a
similar CPU usage in the control plane and will help to generate a more
deterministic load.

Signed-off-by: Raul Sevilla <rsevilla@redhat.com>